Refactor form files
