+++ src/form/addresses/address-utils.ts
@@ -1,0 +1,695 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * OpenCRVS is also distributed under the terms of the Civil Registration
+ * & Healthcare Disclaimer located at http://opencrvs.org/license.
+ *
+ * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
+ * graphic logo are (registered/a) trademark(s) of Plan International.
+ */
+
+import { MessageDescriptor } from 'react-intl'
+import { ADMIN_LEVELS, defaultAddressConfiguration } from '.'
+import {
+  ISerializedForm,
+  SerializedFormField,
+  Event,
+  Conditional,
+  IPreviewGroup
+} from '../types/types'
+import {
+  getAddressFields,
+  getPlaceOfEventAddressFields,
+  getXAddressSameAsY
+} from './address-fields'
+import { getPreviewGroups } from '../common/preview-groups'
+import { cloneDeep } from 'lodash'
+
+export enum AddressCopyConfigCases {
+  PRIMARY_ADDRESS_SAME_AS_OTHER_PRIMARY = 'primaryAddressSameAsOtherPrimary'
+}
+
+export enum AddressSubsections {
+  PRIMARY_ADDRESS_SUBSECTION = 'primaryAddress',
+  SECONDARY_ADDRESS_SUBSECTION = 'secondaryAddress'
+}
+
+export const FATHER_DETAILS_DONT_EXIST =
+  '(draftData?.father && !draftData?.father.detailsExist) || !values.detailsExist'
+export const MOTHER_DETAILS_DONT_EXIST =
+  '(draftData?.mother && !draftData?.mother.detailsExist) || !values.detailsExist'
+
+// if mothers details do not exist on other page
+export const mothersDetailsDontExistOnOtherPage =
+  'draftData && draftData.mother && !draftData.mother.detailsExist'
+
+// if fathers details do not exist
+export const fathersDetailsDontExist = '!values.detailsExist'
+
+// primary address same as other primary
+export const primaryAddressSameAsOtherPrimaryAddress =
+  'values.primaryAddressSameAsOtherPrimary'
+
+export enum EventLocationAddressCases {
+  PLACE_OF_BIRTH = 'placeOfBirth',
+  PLACE_OF_DEATH = 'placeOfDeath',
+  PLACE_OF_MARRIAGE = 'placeOfMarriage'
+}
+
+export enum AddressCases {
+  PRIMARY_ADDRESS = 'PRIMARY_ADDRESS',
+  SECONDARY_ADDRESS = 'SECONDARY_ADDRESS'
+}
+
+export interface IAddressConfiguration {
+  precedingFieldId: string
+  configurations: AllowedAddressConfigurations[]
+}
+
+export type AllowedAddressConfigurations = {
+  config:
+    | AddressCases
+    | AddressSubsections
+    | AddressCopyConfigCases
+    | EventLocationAddressCases
+  label?: MessageDescriptor
+  xComparisonSection?: string
+  yComparisonSection?: string
+  conditionalCase?: string
+}
+
+export function getPlaceOfEventConditionals(
+  location: string,
+  configCase: EventLocationAddressCases
+) {
+  switch (location) {
+    case 'state':
+      return [
+        {
+          action: 'hide',
+          expression: '!values.country'
+        },
+        {
+          action: 'hide',
+          expression:
+            configCase !== EventLocationAddressCases.PLACE_OF_MARRIAGE
+              ? `(values.${configCase}!="OTHER" && values.${configCase}!="PRIVATE_HOME")`
+              : ''
+        },
+        {
+          action: 'hide',
+          expression: '!isDefaultCountry(values.country)'
+        }
+      ]
+    case 'district':
+      return [
+        {
+          action: 'hide',
+          expression: '!values.country'
+        },
+        {
+          action: 'hide',
+          expression: '!values.state'
+        },
+        {
+          action: 'hide',
+          expression:
+            configCase !== EventLocationAddressCases.PLACE_OF_MARRIAGE
+              ? `(values.${configCase}!="OTHER" && values.${configCase}!="PRIVATE_HOME")`
+              : ''
+        },
+        {
+          action: 'hide',
+          expression: '!isDefaultCountry(values.country)'
+        }
+      ]
+    case 'locationLevel3':
+      return [
+        {
+          action: 'hide',
+          expression: '!values.country'
+        },
+        {
+          action: 'hide',
+          expression: '!values.state'
+        },
+        {
+          action: 'hide',
+          expression: '!values.district'
+        },
+        {
+          action: 'hide',
+          expression:
+            configCase !== EventLocationAddressCases.PLACE_OF_MARRIAGE
+              ? `(values.${configCase}!="OTHER" && values.${configCase}!="PRIVATE_HOME")`
+              : ''
+        },
+        {
+          action: 'hide',
+          expression: '!isDefaultCountry(values.country)'
+        }
+      ]
+    case 'locationLevel4':
+      return [
+        {
+          action: 'hide',
+          expression: '!values.country'
+        },
+        {
+          action: 'hide',
+          expression: '!values.state'
+        },
+        {
+          action: 'hide',
+          expression: '!values.district'
+        },
+        {
+          action: 'hide',
+          expression: '!values.locationLevel3'
+        },
+        {
+          action: 'hide',
+          expression:
+            configCase !== EventLocationAddressCases.PLACE_OF_MARRIAGE
+              ? `(values.${configCase}!="OTHER" && values.${configCase}!="PRIVATE_HOME")`
+              : ''
+        },
+        {
+          action: 'hide',
+          expression: '!isDefaultCountry(values.country)'
+        }
+      ]
+    case 'locationLevel5':
+      return [
+        {
+          action: 'hide',
+          expression: '!values.country'
+        },
+        {
+          action: 'hide',
+          expression: '!values.state'
+        },
+        {
+          action: 'hide',
+          expression: '!values.district'
+        },
+        {
+          action: 'hide',
+          expression: '!values.locationLevel3'
+        },
+        {
+          action: 'hide',
+          expression: '!values.locationLevel4'
+        },
+        {
+          action: 'hide',
+          expression:
+            configCase !== EventLocationAddressCases.PLACE_OF_MARRIAGE
+              ? `(values.${configCase}!="OTHER" && values.${configCase}!="PRIVATE_HOME")`
+              : ''
+        },
+        {
+          action: 'hide',
+          expression: '!isDefaultCountry(values.country)'
+        }
+      ]
+    default:
+      throw Error(
+        'Supplied event location is unsupported by current conditionals'
+      )
+  }
+}
+
+export function getRuralOrUrbanConditionals(
+  useCase: string,
+  defaultConditionals: Conditional[]
+) {
+  let customConditionals: Conditional[] = []
+  switch (ADMIN_LEVELS) {
+    case 1:
+      customConditionals = [
+        {
+          action: 'hide',
+          expression: `!values.state${sentenceCase(useCase)}`
+        }
+      ]
+      break
+    case 2:
+      customConditionals = [
+        {
+          action: 'hide',
+          expression: `!values.state${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.district${sentenceCase(useCase)}`
+        }
+      ]
+      break
+    case 3:
+      customConditionals = [
+        {
+          action: 'hide',
+          expression: `!values.state${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.district${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.locationLevel3${sentenceCase(useCase)}`
+        }
+      ]
+      break
+    case 4:
+      customConditionals = [
+        {
+          action: 'hide',
+          expression: `!values.state${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.district${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.locationLevel3${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.locationLevel4${sentenceCase(useCase)}`
+        }
+      ]
+      break
+  }
+  return defaultConditionals.concat(customConditionals)
+}
+
+export function getAddressConditionals(location: string, useCase: string) {
+  switch (location) {
+    case 'state':
+      return [
+        {
+          action: 'hide',
+          expression: `!values.country${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!isDefaultCountry(values.country${sentenceCase(
+            useCase
+          )})`
+        }
+      ]
+    case 'district':
+      return [
+        {
+          action: 'hide',
+          expression: `!values.country${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.state${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!isDefaultCountry(values.country${sentenceCase(
+            useCase
+          )})`
+        }
+      ]
+    case 'locationLevel3':
+      return [
+        {
+          action: 'hide',
+          expression: `!values.country${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.state${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.district${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!isDefaultCountry(values.country${sentenceCase(
+            useCase
+          )})`
+        }
+      ]
+    case 'locationLevel4':
+      return [
+        {
+          action: 'hide',
+          expression: `!values.country${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.state${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.district${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.locationLevel3${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!isDefaultCountry(values.country${sentenceCase(
+            useCase
+          )})`
+        }
+      ]
+    case 'locationLevel5':
+      return [
+        {
+          action: 'hide',
+          expression: `!values.country${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.state${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.district${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.locationLevel3${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!values.locationLevel4${sentenceCase(useCase)}`
+        },
+        {
+          action: 'hide',
+          expression: `!isDefaultCountry(values.country${sentenceCase(
+            useCase
+          )})`
+        }
+      ]
+    default:
+      throw Error('Supplied location is unsupported by current conditionals')
+  }
+}
+export const sentenceCase = (str: string): string =>
+  str.replace(/\w\S*/g, (txt: string) => {
+    return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase()
+  })
+
+export function getDependency(location: string, useCase: string) {
+  switch (location) {
+    case 'state':
+      return useCase === 'placeOfBirth' ||
+        useCase === 'placeOfDeath' ||
+        useCase === 'placeOfMarriage'
+        ? 'country'
+        : `country${sentenceCase(useCase)}`
+    case 'district':
+      return useCase === 'placeOfBirth' ||
+        useCase === 'placeOfDeath' ||
+        useCase === 'placeOfMarriage'
+        ? 'state'
+        : `state${sentenceCase(useCase)}`
+    case 'locationLevel3':
+      return useCase === 'placeOfBirth' ||
+        useCase === 'placeOfDeath' ||
+        useCase === 'placeOfMarriage'
+        ? 'district'
+        : `district${sentenceCase(useCase)}`
+    case 'locationLevel4':
+      return useCase === 'placeOfBirth' ||
+        useCase === 'placeOfDeath' ||
+        useCase === 'placeOfMarriage'
+        ? 'locationLevel3'
+        : `locationLevel3${sentenceCase(useCase)}`
+    case 'locationLevel5':
+      return useCase === 'placeOfBirth' ||
+        useCase === 'placeOfDeath' ||
+        useCase === 'placeOfMarriage'
+        ? 'locationLevel4'
+        : `locationLevel4${sentenceCase(useCase)}`
+    default:
+      throw Error('Supplied address dependency is unsupported')
+  }
+}
+
+export function getCustomAdminLevelSelect(
+  location: string,
+  useCase: string,
+  locationIndex: number
+): SerializedFormField {
+  return {
+    name: `${location}${sentenceCase(useCase)}`,
+    type: 'SELECT_WITH_DYNAMIC_OPTIONS',
+    label: {
+      defaultMessage: '',
+      description: `Title for the ${location} select`,
+      id: `form.field.label.${location}`
+    },
+    previewGroup: `${useCase}Address`,
+    required: true,
+    initialValue: '',
+    validator: [],
+    placeholder: {
+      defaultMessage: 'Select',
+      description: 'Placeholder text for a select',
+      id: 'form.field.select.placeholder'
+    },
+    dynamicOptions: {
+      resource: 'locations',
+      dependency: getDependency(location, useCase),
+      initialValue: 'agentDefault'
+    },
+    conditionals: getAddressConditionals(location, useCase),
+    mapping: {
+      mutation: {
+        operation: 'fieldToAddressLineTransformer',
+        parameters: [
+          useCase.toUpperCase() === 'PRIMARY'
+            ? AddressCases.PRIMARY_ADDRESS
+            : AddressCases.SECONDARY_ADDRESS,
+          locationIndex
+        ]
+      },
+      query: {
+        operation: 'addressLineToFieldTransformer',
+        parameters: [
+          useCase.toUpperCase() === 'PRIMARY'
+            ? AddressCases.PRIMARY_ADDRESS
+            : AddressCases.SECONDARY_ADDRESS,
+          locationIndex,
+          location
+        ]
+      }
+    }
+  }
+}
+
+export function getIdentifiersFromFieldId(fieldId: string) {
+  const splitIds = fieldId.split('.')
+  return {
+    event: splitIds[0] as Event,
+    sectionId: splitIds[1],
+    groupId: splitIds[2],
+    fieldName: splitIds[3]
+  }
+}
+
+export function getSectionIdentifiers(fieldId: string, form: ISerializedForm) {
+  const { event, sectionId } = getIdentifiersFromFieldId(fieldId)
+
+  const sectionIndex = form.sections.findIndex(({ id }) => id === sectionId)
+  return {
+    event,
+    sectionIndex
+  }
+}
+
+export function getGroupIdentifiers(fieldId: string, form: ISerializedForm) {
+  const { event, groupId } = getIdentifiersFromFieldId(fieldId)
+
+  const { sectionIndex } = getSectionIdentifiers(fieldId, form)
+
+  const groups = form.sections[sectionIndex].groups
+
+  const groupIndex = groups.findIndex(({ id }) => id === groupId)
+
+  return {
+    event,
+    sectionIndex,
+    groupIndex
+  }
+}
+
+export function getFieldIdentifiers(fieldId: string, form: ISerializedForm) {
+  const { event, fieldName } = getIdentifiersFromFieldId(fieldId)
+
+  const { sectionIndex, groupIndex } = getGroupIdentifiers(fieldId, form)
+
+  const fields = form.sections[sectionIndex].groups[groupIndex].fields
+
+  const fieldIndex = fields.findIndex(({ name }) => name === fieldName)
+
+  return {
+    event,
+    sectionIndex,
+    groupIndex,
+    fieldIndex
+  }
+}
+
+export const getAddressSubsection = (
+  previewGroup: AddressSubsections,
+  label: MessageDescriptor,
+  conditionalCase?: string
+): SerializedFormField[] => {
+  const fields: SerializedFormField[] = []
+  const subsection: SerializedFormField = {
+    name: previewGroup,
+    type: 'SUBSECTION_HEADER',
+    label,
+    previewGroup: previewGroup,
+    initialValue: '',
+    validator: []
+  }
+
+  if (conditionalCase) {
+    subsection['conditionals'] = [
+      {
+        action: 'hide',
+        expression: `${conditionalCase}`
+      }
+    ]
+  }
+  fields.push(subsection)
+  return fields
+}
+
+function getAddressFieldsByConfiguration(
+  configuration: AllowedAddressConfigurations
+): SerializedFormField[] {
+  switch (configuration.config) {
+    case EventLocationAddressCases.PLACE_OF_BIRTH:
+    case EventLocationAddressCases.PLACE_OF_DEATH:
+    case EventLocationAddressCases.PLACE_OF_MARRIAGE:
+      return getPlaceOfEventAddressFields(configuration.config)
+    case AddressCases.PRIMARY_ADDRESS:
+      return getAddress(
+        AddressCases.PRIMARY_ADDRESS,
+        configuration.conditionalCase
+      )
+    case AddressCases.SECONDARY_ADDRESS:
+      return getAddress(
+        AddressCases.SECONDARY_ADDRESS,
+        configuration.conditionalCase
+      )
+    case AddressCopyConfigCases.PRIMARY_ADDRESS_SAME_AS_OTHER_PRIMARY:
+      if (
+        !configuration.label ||
+        !configuration.xComparisonSection ||
+        !configuration.yComparisonSection
+      ) {
+        throw new Error(
+          `Invalid address comparison case configuration for: ${configuration.config}`
+        )
+      }
+      return getXAddressSameAsY(
+        configuration.xComparisonSection,
+        configuration.yComparisonSection,
+        configuration.label,
+        configuration.conditionalCase
+      )
+    case AddressSubsections.PRIMARY_ADDRESS_SUBSECTION:
+    case AddressSubsections.SECONDARY_ADDRESS_SUBSECTION:
+      if (!configuration.label) {
+        throw new Error(
+          `Invalid address subsection configuration for: ${configuration.config}`
+        )
+      }
+      return getAddressSubsection(
+        configuration.config,
+        configuration.label,
+        configuration.conditionalCase
+      )
+    default:
+      return []
+  }
+}
+
+export function decorateFormsWithAddresses(
+  defaultEventForm: ISerializedForm,
+  event: string
+): ISerializedForm {
+  const newForm = cloneDeep(defaultEventForm)
+  defaultAddressConfiguration.forEach(
+    ({ precedingFieldId, configurations }: IAddressConfiguration) => {
+      if (precedingFieldId.includes(event)) {
+        const { sectionIndex, groupIndex, fieldIndex } = getFieldIdentifiers(
+          precedingFieldId,
+          newForm
+        )
+
+        let addressFields: SerializedFormField[] = []
+        let previewGroups: IPreviewGroup[] = []
+        configurations.forEach((configuration) => {
+          addressFields = addressFields.concat(
+            getAddressFieldsByConfiguration(configuration)
+          )
+          previewGroups = previewGroups.concat(getPreviewGroups(configuration))
+        })
+        newForm.sections[sectionIndex].groups[groupIndex].fields.splice(
+          fieldIndex + 1,
+          0,
+          ...addressFields
+        )
+
+        const group = newForm.sections[sectionIndex].groups[groupIndex]
+        if (group.previewGroups) {
+          group.previewGroups = group.previewGroups.concat(previewGroups)
+        } else {
+          group.previewGroups = previewGroups
+        }
+      }
+    }
+  )
+
+  return newForm
+}
+
+export function getAddress(
+  addressCase: AddressCases,
+  conditionalCase?: string
+): SerializedFormField[] {
+  const defaultFields: SerializedFormField[] = getAddressFields(addressCase)
+  if (conditionalCase) {
+    defaultFields.forEach((field) => {
+      let conditional
+      if (conditionalCase) {
+        conditional = {
+          action: 'hide',
+          expression: `${conditionalCase}`
+        }
+      }
+      if (
+        conditional &&
+        field.conditionals &&
+        field.conditionals.filter(
+          (conditional) => conditional.expression === conditionalCase
+        ).length === 0
+      ) {
+        field.conditionals.push(conditional)
+      } else if (conditional && !field.conditionals) {
+        field.conditionals = [conditional]
+      }
+    })
+  }
+  return defaultFields
+}
