+++ src/form/birth/certificate-handlebars.ts
@@ -1,0 +1,62 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * OpenCRVS is also distributed under the terms of the Civil Registration
+ * & Healthcare Disclaimer located at http://opencrvs.org/license.
+ *
+ * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
+ * graphic logo are (registered/a) trademark(s) of Plan International.
+ */
+
+// THE FOLLOWING HANDLEBARS CAN BE USED IN THE SVG CODE IN THE CERTIFICATE FOR THIS EVENT
+
+export const certificateHandlebars = {
+  registrationNumber: 'registrationNumber',
+  qrCode: 'qrCode',
+  mosipAid: 'mosipAid',
+  mosipAIDLabel: 'mosipAIDLabel',
+  certificateDate: 'certificateDate',
+  registrar: 'registrar',
+  registrationAgent: 'registrationAgent',
+  registrarName: 'registrarName',
+  role: 'role',
+  registrarSignature: 'registrarSignature',
+  registrationDate: 'registrationDate',
+  registrationLocation: 'registrationLocation',
+  contactEmail: 'contactEmail',
+  contactPhoneNumber: 'contactPhoneNumber',
+  placeOfBirth: 'placeOfBirth',
+  informantType: 'informantType',
+  attendantAtBirth: 'attendantAtBirth',
+  multipleBirth: 'multipleBirth',
+  weightAtBirth: 'weightAtBirth',
+  birthType: 'birthType',
+  childFirstName: 'childFirstName',
+  childFamilyName: 'childFamilyName',
+  childGender: 'childGender',
+  eventDate: 'eventDate',
+  informantFirstName: 'informantFirstName',
+  informantFamilyName: 'informantFamilyName',
+  informantBirthDate: 'informantBirthDate',
+  informantNationality: 'informantNationality',
+  informantNID: 'informantNID',
+  motherReasonNotApplying: 'motherReasonNotApplying',
+  motherBirthDate: 'motherBirthDate',
+  motherFirstName: 'motherFirstName',
+  motherFamilyName: 'motherFamilyName',
+  motherNationality: 'motherNationality',
+  motherNID: 'motherNID',
+  motherMaritalStatus: 'motherMaritalStatus',
+  motherOccupation: 'motherOccupation',
+  motherEducationalAttainment: 'motherEducationalAttainment',
+  fatherBirthDate: 'fatherBirthDate',
+  fatherFirstName: 'fatherFirstName',
+  fatherFamilyName: 'fatherFamilyName',
+  fatherNationality: 'fatherNationality',
+  fatherNID: 'fatherNID',
+  fatherMaritalStatus: 'fatherMaritalStatus',
+  fatherOccupation: 'fatherOccupation',
+  fatherEducationalAttainment: 'fatherEducationalAttainment'
+}
