+++ src/form/birth/required-sections.ts
@@ -1,0 +1,306 @@
+import { formMessageDescriptors } from '../common/messages'
+import { ISerializedFormSection } from '../types/types'
+import { certificateHandlebars } from './certificate-handlebars'
+
+export const registrationSection = {
+  id: 'registration', // A hidden 'registration' section must be included to store identifiers in a form draft that are used in certificates
+  viewType: 'form',
+  name: {
+    defaultMessage: 'Registration',
+    description: 'Form section name for Registration',
+    id: 'form.section.declaration.name'
+  },
+  groups: [],
+  mapping: {
+    template: [
+      {
+        fieldName: certificateHandlebars.registrationNumber,
+        operation: 'registrationNumberTransformer'
+      },
+      {
+        fieldName: certificateHandlebars.qrCode,
+        operation: 'QRCodeTransformerTransformer'
+      },
+      {
+        fieldName: certificateHandlebars.mosipAid,
+        operation: 'mosipAidTransformer'
+      },
+      {
+        fieldName: certificateHandlebars.mosipAIDLabel,
+        operation: 'mosipAidLabelTransformer'
+      },
+      {
+        fieldName: certificateHandlebars.certificateDate,
+        operation: 'certificateDateTransformer',
+        parameters: ['en', 'dd MMMM yyyy']
+      },
+      {
+        fieldName: certificateHandlebars.registrar,
+        operation: 'userTransformer',
+        parameters: ['REGISTERED']
+      },
+      {
+        fieldName: certificateHandlebars.registrationAgent,
+        operation: 'userTransformer',
+        parameters: ['VALIDATED']
+      },
+      {
+        fieldName: certificateHandlebars.registrarName,
+        operation: 'registrarNameUserTransformer'
+      },
+      {
+        fieldName: certificateHandlebars.role,
+        operation: 'roleUserTransformer'
+      },
+      {
+        fieldName: certificateHandlebars.registrarSignature,
+        operation: 'registrarSignatureUserTransformer'
+      },
+      {
+        fieldName: certificateHandlebars.registrationDate,
+        operation: 'registrationDateTransformer',
+        parameters: ['en', 'dd MMMM yyyy']
+      },
+      {
+        fieldName: certificateHandlebars.registrationLocation,
+        operation: 'registrationLocationUserTransformer'
+      }
+    ],
+    mutation: {
+      operation: 'setBirthRegistrationSectionTransformer'
+    },
+    query: {
+      operation: 'getBirthRegistrationSectionTransformer'
+    }
+  }
+} as ISerializedFormSection
+
+export const birthDocumentExtraValue = {
+  CHILD: 'CHILD',
+  FATHER: 'FATHER',
+  MOTHER: 'MOTHER',
+  PARENT: 'PARENT',
+  OTHER: 'OTHER',
+  INFORMANT_ID_PROOF: 'INFORMANT_ID_PROOF',
+  LEGAL_GUARDIAN_PROOF: 'LEGAL_GUARDIAN_PROOF'
+}
+
+export const birthDocumentType = {
+  BIRTH_CERTIFICATE: 'BIRTH_CERTIFICATE',
+  NATIONAL_ID: 'NATIONAL_ID',
+  PASSPORT: 'PASSPORT',
+  OTHER: 'OTHER',
+  NOTIFICATION_OF_BIRTH: 'NOTIFICATION_OF_BIRTH',
+  PROOF_OF_LEGAL_GUARDIANSHIP: 'PROOF_OF_LEGAL_GUARDIANSHIP',
+  PROOF_OF_ASSIGNED_RESPONSIBILITY: 'PROOF_OF_ASSIGNED_RESPONSIBILITY'
+}
+
+export const documentsSection = {
+  id: 'documents',
+  viewType: 'form',
+  name: formMessageDescriptors.documentsName,
+  title: {
+    defaultMessage: 'Attaching supporting documents',
+    description: 'Form section title for Documents',
+    id: 'form.section.documents.title'
+  },
+  groups: [
+    {
+      id: 'documents-view-group',
+      fields: [
+        {
+          name: 'paragraph',
+          type: 'PARAGRAPH',
+          label: formMessageDescriptors.documentsParagraph,
+          initialValue: '',
+          validator: []
+        },
+        {
+          name: 'uploadDocForChildDOB',
+          type: 'DOCUMENT_UPLOADER_WITH_OPTION',
+          label: formMessageDescriptors.proofOfBirth,
+          initialValue: '',
+          extraValue: birthDocumentExtraValue.CHILD,
+          hideAsterisk: true,
+          validator: [],
+          options: [
+            {
+              value: birthDocumentType.NOTIFICATION_OF_BIRTH,
+              label: formMessageDescriptors.docTypeChildBirthProof
+            }
+          ],
+          mapping: {
+            mutation: {
+              operation: 'eventFieldToAttachmentTransformer'
+            },
+            query: {
+              operation: 'eventAttachmentToFieldTransformer'
+            }
+          }
+        },
+        {
+          name: 'uploadDocForMother',
+          type: 'DOCUMENT_UPLOADER_WITH_OPTION',
+          label: formMessageDescriptors.proofOfMothersID,
+          initialValue: '',
+          extraValue: birthDocumentExtraValue.MOTHER,
+          hideAsterisk: true,
+          validator: [],
+          options: [
+            {
+              value: birthDocumentType.NATIONAL_ID,
+              label: formMessageDescriptors.docTypeNID
+            },
+            {
+              value: birthDocumentType.PASSPORT,
+              label: formMessageDescriptors.docTypePassport
+            },
+            {
+              value: birthDocumentType.BIRTH_CERTIFICATE,
+              label: formMessageDescriptors.docTypeBirthCert
+            },
+            {
+              value: birthDocumentType.OTHER,
+              label: formMessageDescriptors.docTypeOther
+            }
+          ],
+          conditionals: [
+            {
+              description: 'Hidden for Parent Details none or Mother only',
+              action: 'hide',
+              expression:
+                'draftData && draftData.mother && !draftData.mother.detailsExist'
+            }
+          ],
+          mapping: {
+            mutation: {
+              operation: 'eventFieldToAttachmentTransformer'
+            },
+            query: {
+              operation: 'eventAttachmentToFieldTransformer'
+            }
+          }
+        },
+        {
+          name: 'uploadDocForFather',
+          type: 'DOCUMENT_UPLOADER_WITH_OPTION',
+          label: formMessageDescriptors.proofOfFathersID,
+          initialValue: '',
+          extraValue: birthDocumentExtraValue.FATHER,
+          hideAsterisk: true,
+          validator: [],
+          options: [
+            {
+              value: birthDocumentType.NATIONAL_ID,
+              label: formMessageDescriptors.docTypeNID
+            },
+            {
+              value: birthDocumentType.PASSPORT,
+              label: formMessageDescriptors.docTypePassport
+            },
+            {
+              value: birthDocumentType.BIRTH_CERTIFICATE,
+              label: formMessageDescriptors.docTypeBirthCert
+            },
+            {
+              value: birthDocumentType.OTHER,
+              label: formMessageDescriptors.docTypeOther
+            }
+          ],
+          conditionals: [
+            {
+              description: 'Hidden for Parent Details none or Father only',
+              action: 'hide',
+              expression:
+                'draftData && draftData.father && !draftData.father.detailsExist'
+            }
+          ],
+          mapping: {
+            mutation: {
+              operation: 'eventFieldToAttachmentTransformer'
+            },
+            query: {
+              operation: 'eventAttachmentToFieldTransformer'
+            }
+          }
+        },
+        {
+          name: 'uploadDocForInformant',
+          type: 'DOCUMENT_UPLOADER_WITH_OPTION',
+          label: formMessageDescriptors.proofOfInformantsID,
+          initialValue: '',
+          extraValue: birthDocumentExtraValue.INFORMANT_ID_PROOF,
+          hideAsterisk: true,
+          validator: [],
+          options: [
+            {
+              value: birthDocumentType.NATIONAL_ID,
+              label: formMessageDescriptors.docTypeNID
+            },
+            {
+              value: birthDocumentType.PASSPORT,
+              label: formMessageDescriptors.docTypePassport
+            },
+            {
+              value: birthDocumentType.BIRTH_CERTIFICATE,
+              label: formMessageDescriptors.docTypeBirthCert
+            },
+            {
+              value: birthDocumentType.OTHER,
+              label: formMessageDescriptors.docTypeOther
+            }
+          ],
+          conditionals: [
+            {
+              action: 'hide',
+              expression:
+                "draftData?.informant?.informantType === 'MOTHER' || draftData?.informant?.informantType === 'FATHER'"
+            }
+          ],
+          mapping: {
+            mutation: {
+              operation: 'eventFieldToAttachmentTransformer'
+            },
+            query: {
+              operation: 'eventAttachmentToFieldTransformer'
+            }
+          }
+        },
+        {
+          name: 'uploadDocForProofOfLegalGuardian',
+          type: 'DOCUMENT_UPLOADER_WITH_OPTION',
+          label: formMessageDescriptors.otherBirthSupportingDocuments,
+          initialValue: '',
+          extraValue: birthDocumentExtraValue.LEGAL_GUARDIAN_PROOF,
+          hideAsterisk: true,
+          validator: [],
+          options: [
+            {
+              value: birthDocumentType.PROOF_OF_LEGAL_GUARDIANSHIP,
+              label: formMessageDescriptors.legalGuardianProof
+            },
+            {
+              value: birthDocumentType.PROOF_OF_ASSIGNED_RESPONSIBILITY,
+              label: formMessageDescriptors.assignedResponsibilityProof
+            }
+          ],
+          conditionals: [
+            {
+              action: 'hide',
+              expression:
+                "(draftData && draftData.registration && draftData.registration.informantType && selectedInformantAndContactType.selectedInformantType && (selectedInformantAndContactType.selectedInformantType === 'MOTHER' || selectedInformantAndContactType.selectedInformantType === 'FATHER'))"
+            }
+          ],
+          mapping: {
+            mutation: {
+              operation: 'eventFieldToAttachmentTransformer'
+            },
+            query: {
+              operation: 'eventAttachmentToFieldTransformer'
+            }
+          }
+        }
+      ]
+    }
+  ]
+} as ISerializedFormSection
