+++ src/form/death/certficate-handlebars.ts
@@ -1,0 +1,47 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * OpenCRVS is also distributed under the terms of the Civil Registration
+ * & Healthcare Disclaimer located at http://opencrvs.org/license.
+ *
+ * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
+ * graphic logo are (registered/a) trademark(s) of Plan International.
+ */
+
+// THE FOLLOWING HANDLEBARS CAN BE USED IN THE SVG CODE IN THE CERTIFICATE FOR THIS EVENT
+
+export const certificateHandlebars = {
+  registrationNumber: 'registrationNumber',
+  qrCode: 'qrCode',
+  certificateDate: 'certificateDate',
+  registrar: 'registrar',
+  registrationAgent: 'registrationAgent',
+  registrarName: 'registrarName',
+  role: 'role',
+  registrarSignature: 'registrarSignature',
+  registrationDate: 'registrationDate',
+  registrationLocation: 'registrationLocation',
+  contactEmail: 'contactEmail',
+  contactPhoneNumber: 'contactPhoneNumber',
+  placeOfDeath: 'placeOfDeath',
+  deathDescription: 'deathDescription',
+  causeOfDeathMethod: 'causeOfDeathMethod',
+  causeOfDeathEstablished: 'causeOfDeathEstablished',
+  mannerOfDeath: 'mannerOfDeath',
+  informantType: 'informantType',
+  eventDate: 'eventDate',
+  informantNID: 'informantNID',
+  informantNationality: 'informantNationality',
+  informantBirthDate: 'informantBirthDate',
+  informantFamilyName: 'informantFamilyName',
+  informantFirstName: 'informantFirstName',
+  deceasedMaritalStatus: 'deceasedMaritalStatus',
+  deceasedNID: 'deceasedNID',
+  deceasedNationality: 'deceasedNationality',
+  deceasedBirthDate: 'deceasedBirthDate',
+  deceasedGender: 'deceasedGender',
+  deceasedFirstName: 'deceasedFirstName',
+  deceasedFamilyName: 'deceasedFamilyName'
+}
