+++ src/form/index.ts
@@ -10,39 +10,50 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 
-import { birthRegisterForms as birth } from './birth'
-import { deathRegisterForms as death } from './death'
-import { marriageRegisterForms as marriage } from './marriage'
-import { populateRegisterFormsWithAddresses } from './addresses'
+import { decorateFormsWithAddresses } from './addresses/address-utils'
+import { birthForm } from './birth/birth'
+import { deathForm } from './death/death'
+import { marriageForm } from './marriage/marriage'
 import { IForms, Event } from './types/types'
 
 export async function formHandler(): Promise<IForms> {
-  // AS OF OPENCRVS v1.3.0 YOU CANNOT MAKE A CHANGE TO A CIVIL REGISTRATION FORM ONCE YOU GO LIVE IN PRODUCTION
+  // ====================== NOTE REGARDING MIGRATING FROM OPNCRVS v1.2 OR EARLIER ======================
 
-  // AFTER YOU HAVE STARTED REGISTERING CITIZENS, WHEN YOU MAKE A CHANGE TO A FORM IN PRODUCTION,
-  // THE FORM MAY NO LONGER WORK FOR REGISTRATIONS MADE ON A PREVIOUS VERSION OF THE FORM.
-
-  // THIS IS BECAUSE, YOU MAY HAVE REMOVED PREVIOUSLY REQUIRED QUESTIONS, OR ADDED NEW ONES.
-
-  // AN OLDER REGISTRATION MAY NOT HAVE CAPTURED SUCH DATA, SO IF YOU VIEW OR CORRECT THAT RECORD
-  // YOU MAY FIND THAT REQUIRED DATA IS NOW MISSING AND FORM VALIDATION PREVENTS THE SUBMISSION OF THE FORM.
-
-  // WE HAVE A BACKLOG ITEM TO ALLOW REGISTRATION FORM CHANGES TO BE MADE POST LIVE TO
-  // SUPPORT CIVIL REGISTRATION REGULATORY CHANGES THAT MAY HAPPEN IN THE FUTURE.
-
-  // FOLLOW THE STATUS OF THIS ISSUE: https://github.com/opencrvs/opencrvs-core/issues/3798
-
-  // IN THE MEANTIME, WE WILL SAVE A VERSION OF YOUR FORM IN AN IMMUTABLE DATABASE, WHENEVER YOU EDIT THE VERSION NUMBER BELOW
-  // THIS WILL ALLOW US IN FUTURE TO ASSOCIATE A REGISTRATION WITH A PREVIOUSLY DRAFTED REGISTRATION FORM
-  // EVEN IF THE CIVIL REGISTRATION REGULATIONS EVOLVE IN YOUR COUNTRY.
+  // SIMPLY RETURN A JSON OF YOUR FULL FORM HERE, WITH THE ADDITION OF THE NEW MARRIAGE AND VERSION PROP
+  // ALTERNATIVELY, REFACTOR YOUR JSON INTO TYPESCRIPT FOLLOWING OUR EXAMPLE
 
   return {
-    // THE VERSION NUMBER FOR YOUR FORM IS NOT RELATED TO ANY OPENCRVS VERSION.
-    // YOU CAN NAME THIS VERSION ANYTHING YOU LIKE AS LONG AS IT IS UNIQUE EACH TIME.
+    // ====================== NOTE REGARDING FORM VERSION ======================
+
+    // THE VERSION NUMBER FOR YOUR FORM IS NOT RELATED TO THE OPENCRVS VERSION.
+    // THIS IS A VERSION NUMBER FOR YOUR USE AS YOU EDIT FORMS OVER TIME.
+
+    // AFTER YOU HAVE STARTED REGISTERING CITIZENS, WHEN YOU MAKE A CHANGE TO A FORM IN PRODUCTION,
+    // THE FORM MAY NO LONGER WORK FOR REGISTRATIONS MADE ON A PREVIOUS VERSION OF THE FORM.
+
+    // THIS IS BECAUSE, YOU MAY HAVE REMOVED PREVIOUSLY REQUIRED QUESTIONS, OR ADDED NEW ONES.
+
+    // AN OLDER REGISTRATION MAY NOT HAVE CAPTURED SUCH DATA, SO IF YOU VIEW OR CORRECT THAT RECORD
+    // YOU MAY FIND THAT REQUIRED DATA IS NOW MISSING AND FORM VALIDATION PREVENTS THE SUBMISSION OF THE FORM.
+
+    // WE HAVE A BACKLOG ITEM TO ALLOW REGISTRATION FORM CHANGES TO BE MADE POST LIVE TO
+    // SUPPORT CIVIL REGISTRATION REGULATORY CHANGES THAT MAY HAPPEN IN THE FUTURE.
+
+    // FOLLOW THE STATUS OF THIS ISSUE: https://github.com/opencrvs/opencrvs-core/issues/3798
+
+    // IN THE MEANTIME, WHENEVER YOU SEMANTICALLY INCREMENT THE VERSION NUMBER BELOW, WE WILL SAVE A VERSION OF YOUR FORM IN AN IMMUTABLE DATABASE.
+
+    // THIS WILL ALLOW US IN FUTURE TO ASSOCIATE A REGISTRATION WITH A PREVIOUSLY DRAFTED REGISTRATION FORM
+    // EVEN IF THE CIVIL REGISTRATION REGULATIONS EVOLVE IN YOUR COUNTRY.
 
     version: 'v1.0.0',
-    birth: populateRegisterFormsWithAddresses(birth, Event.Birth),
-    death: populateRegisterFormsWithAddresses(death, Event.Death),
-    marriage: populateRegisterFormsWithAddresses(marriage, Event.Marriage)
+
+    // ADDRESS FORMAT IS DYNAMICALLY DUPLICATED MULTIPLE TIMES FOR ALL EVENTS.
+    // THIS DECORATOR FUNCTION POPULATES ADDRESSES ACCORDING TO THE defaultAddressConfiguration in address-settings.ts
+    // SO YOU ONLY NEED TO CONFIGURE ADDRESS FIELDS IN A SINGLE LOCATION FOR ALL DECORATED INSTANCES.
+
+    birth: decorateFormsWithAddresses(birthForm, Event.Birth),
+    death: decorateFormsWithAddresses(deathForm, Event.Death),
+    marriage: decorateFormsWithAddresses(marriageForm, Event.Marriage)
   }
 }
