+++ src/form/marriage/certificate-handlebars.ts
@@ -1,0 +1,57 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * OpenCRVS is also distributed under the terms of the Civil Registration
+ * & Healthcare Disclaimer located at http://opencrvs.org/license.
+ *
+ * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
+ * graphic logo are (registered/a) trademark(s) of Plan International.
+ */
+
+// THE FOLLOWING HANDLEBARS CAN BE USED IN THE SVG CODE IN THE CERTIFICATE FOR THIS EVENT
+
+export const certificateHandlebars = {
+  registrationNumber: 'registrationNumber',
+  qrCode: 'qrCode',
+  certificateDate: 'certificateDate',
+  registrar: 'registrar',
+  registrationAgent: 'registrationAgent',
+  contactEmail: 'contactEmail',
+  contactPhoneNumber: 'contactPhoneNumber',
+  registrarName: 'registrarName',
+  role: 'role',
+  registrarSignature: 'registrarSignature',
+  registrationDate: 'registrationDate',
+  registrationLocation: 'registrationLocation',
+  groomSignature: 'groomSignature',
+  brideSignature: 'brideSignature',
+  witnessOneSignature: 'witnessOneSignature',
+  witnessTwoSignature: 'witnessTwoSignature',
+  informantFirstName: 'informantFirstName',
+  informantFamilyName: 'informantFamilyName',
+  groomFirstName: 'groomFirstName',
+  groomFamilyName: 'groomFamilyName',
+  groomMarriedLastNameEng: 'groomMarriedLastNameEng',
+  brideFirstName: 'brideFirstName',
+  brideFamilyName: 'brideFamilyName',
+  brideMarriedLastNameEng: 'brideMarriedLastNameEng',
+  witnessOneFirstName: 'witnessOneFirstName',
+  witnessOneFamilyName: 'witnessOneFamilyName',
+  witnessTwoFirstName: 'witnessTwoFirstName',
+  witnessTwoFamilyName: 'witnessTwoFamilyName',
+  marriedLastNameEng: 'marriedLastNameEng',
+  eventDate: 'eventDate',
+  informantType: 'informantType',
+  typeOfMarriage: 'typeOfMarriage',
+  informantNID: 'informantNID',
+  informantNationality: 'informantNationality',
+  informantBirthDate: 'informantBirthDate',
+  brideNID: 'brideNID',
+  brideNationality: 'brideNationality',
+  brideBirthDate: 'brideBirthDate',
+  groomNID: 'groomNID',
+  groomNationality: 'groomNationality',
+  groomBirthDate: 'groomBirthDate'
+}
